RavenDB-18729 moving files around
