+++ src/Raven.Server/Documents/ETL/Providers/Raven/Handlers/RavenEtlHandler.cs
@@ -1,13 +1,6 @@
-﻿using System.Linq;
-using System.Threading.Tasks;
-using Raven.Client.Documents.Conventions;
-using Raven.Server.Documents.ETL.Providers.Raven.Test;
-using Raven.Server.Documents.Handlers.Processors.ETL;
-using Raven.Server.Json;
+﻿using System.Threading.Tasks;
+using Raven.Server.Documents.ETL.Providers.Raven.Handlers.Processors;
 using Raven.Server.Routing;
-using Raven.Server.ServerWide.Context;
-using Sparrow.Json;
-using Sparrow.Json.Parsing;
 
 namespace Raven.Server.Documents.ETL.Providers.Raven.Handlers
 {
